refactor(tern): split local_apply.go by apply mode
Move the 1,450-line pkg/tern/local_apply.go into four focused files
within the same package. Pure file-organization refactor — no logic,
signature, or visibility changes.

- local_apply.go: orchestration, dispatcher, event/state helpers
  (checkActiveTaskConflict, runApplyExecution, runWithRecovery,
  logApplyEvent, setupSpiritLogging, transitionTaskState,
  deriveApplyPhase, applyEventStateTransition, deriveOverallState,
  planNamespacesToChanges, ...)
- local_apply_grouped.go: grouped/atomic apply path
  (executeGroupedApply, pollForCompletionAtomic,
  handleAtomicProgressTick, revert-window helpers,
  logAtomicProgress, syncAtomicTaskProgress, deriveVSchemaTaskState)
- local_apply_sequential.go: sequential apply path
  (executeApplySequential, checkTaskReady, runEngineTask,
  startApplyHeartbeat, pollTaskToCompletion, markTaskFailed,
  markTaskRetryable, shouldRetryEngineError, finalizeSequentialApply)
- local_apply_failure.go: cross-path failure escalation
  (failApplyWithTasks, markApplyRetryableWithTasks)

Validation:
- go build ./...
- go test -count=1 -race ./pkg/tern/...
- make test-unit (all packages pass)
- golangci-lint run --new-from-rev=origin/main ./pkg/tern/... (0 issues)
